+++ contracts/utils/ILinearTokenTimelock.sol
@@ -13,7 +13,7 @@ interface ILinearTokenTimelock {
 
     // ----------- State changing api -----------
 
-    function release(address to, uint amount) external;
+    function release(address to, uint256 amount) external;
 
     function releaseMax(address to) external;
